+++ src/num_complex/v04/bnum_0_13.rs
@@ -1,6 +1,73 @@
-use crate::{bnum::Packable, DecodeError, EncodeError, Varint};
+use crate::{
+  bnum::{decode_uint_d8, encode_uint_d8_to, encoded_uint_d8_len, Packable},
+  utils::{pack_i128, pack_u128, unpack_i128, unpack_u128},
+  DecodeError, EncodeError, Varint,
+};
 use ::bnum_0_13::{BInt, BIntD16, BIntD32, BIntD8, BUint, BUintD16, BUintD32, BUintD8};
+
 use num_complex_0_4::Complex;
+
+type U256 = BUintD8<32>;
+
+impl_varint_for_complex!(128(U256));
+
+/// Returns the encoded length of the `Complex<u128>` value.
+#[inline]
+pub const fn encoded_complex_u128_len(val: &Complex<u128>) -> usize {
+  encoded_uint_d8_len(&pack_u128(val.re, val.im))
+}
+
+/// Encodes the `Complex<u128>` value.
+#[inline]
+pub const fn encode_complex_u128(
+  val: &Complex<u128>,
+  buf: &mut [u8],
+) -> Result<usize, EncodeError> {
+  encode_uint_d8_to(pack_u128(val.re, val.im), buf)
+}
+
+/// Decodes the `Complex<u128>` from the given buffer
+///
+/// Returns the bytes read and the value.
+#[inline]
+pub const fn decode_complex_u128(buf: &[u8]) -> Result<(usize, Complex<u128>), DecodeError> {
+  match decode_uint_d8::<32>(buf) {
+    Ok((read, val)) => {
+      let (re, im) = unpack_u128(val);
+      Ok((read, Complex { re, im }))
+    }
+    Err(e) => Err(e),
+  }
+}
+
+/// Returns the encoded length of the `Complex<i128>` value.
+#[inline]
+pub const fn encoded_complex_i128_len(val: &Complex<i128>) -> usize {
+  encoded_uint_d8_len(&pack_i128(val.re, val.im))
+}
+
+/// Encodes the `Complex<i128>` value.
+#[inline]
+pub const fn encode_complex_i128(
+  val: &Complex<i128>,
+  buf: &mut [u8],
+) -> Result<usize, EncodeError> {
+  encode_uint_d8_to(pack_i128(val.re, val.im), buf)
+}
+
+/// Decodes the `Complex<i128>` from the given buffer
+///
+/// Returns the bytes read and the value.
+#[inline]
+pub const fn decode_complex_i128(buf: &[u8]) -> Result<(usize, Complex<i128>), DecodeError> {
+  match decode_uint_d8::<32>(buf) {
+    Ok((read, val)) => {
+      let (re, im) = unpack_i128(val);
+      Ok((read, Complex { re, im }))
+    }
+    Err(e) => Err(e),
+  }
+}
 
 macro_rules! impl_varint_for_complex_bnum {
   (@unsigned $($base:ident($($bytes:literal),+$(,)?)), +$(,)?) => {