+++ src/num_rational/v04.rs
@@ -38,31 +38,36 @@ macro_rules! impl_varint_for_ratio {
     impl_varint_for_ratio!(@inner $(u::$bits($merged_ty)),+);
     impl_varint_for_ratio!(@inner $(i::$bits($merged_ty)),+);
   };
-  (@const_inner $($sign:ident::$bits:literal($merged_ty:ident)),+$(,)?) => {
+  (@const_inner $($sign:ident::$bits:literal($merged_ty:ident) {
+    encoded_len_fn:$encoded_len_fn:path,
+    encode_fn:$encode_fn:path,
+    encode_to_fn:$encode_to_fn:path,
+    decode_fn:$decode_fn:path,
+  }),+$(,)?) => {
     paste::paste! {
       $(
         #[doc = "Returns the encoded length of the `Ratio<" $sign $bits ">` value."]
         #[inline]
         pub const fn [< encoded_ratio_ $sign $bits _len>](val: Ratio<[< $sign $bits >]>) -> usize {
-          $crate::[< encoded_ $merged_ty _varint_len >] ($crate::utils::[< pack_ $sign $bits >](*val.numer(), *val.denom()))
+          $encoded_len_fn($crate::utils::[< pack_ $sign $bits >](*val.numer(), *val.denom()))
         }
 
         #[doc = "Encodes the `Ratio<" $sign $bits ">` value."]
         #[inline]
         pub const fn [< encode_ratio_ $sign $bits >](val: Ratio<[< $sign $bits >]>) -> $crate::[< $merged_ty:camel VarintBuffer>] {
-          $crate::[< encode_ $merged_ty _varint>] ($crate::utils::[< pack_ $sign $bits >](*val.numer(), *val.denom()))
+          $encode_fn ($crate::utils::[< pack_ $sign $bits >](*val.numer(), *val.denom()))
         }
 
         #[doc = "Encodes the `Ratio<" $sign $bits ">` value into the provided buffer."]
         #[inline]
         pub const fn [< encode_ratio_ $sign $bits _to >](val: Ratio<[< $sign $bits >]>, buf: &mut [u8]) -> Result<usize, EncodeError> {
-          $crate::[< encode_ $merged_ty _varint_to>] ($crate::utils::[< pack_ $sign $bits >](*val.numer(), *val.denom()), buf)
+          $encode_to_fn ($crate::utils::[< pack_ $sign $bits >](*val.numer(), *val.denom()), buf)
         }
 
         #[doc = "Decodes the `Ratio<" $sign $bits ">` value from the provided buffer."]
         #[inline]
         pub const fn [< decode_ratio_ $sign $bits >](buf: &[u8]) -> Result<(usize, Ratio<[< $sign $bits >]>), DecodeError> {
-          match $crate::[< decode_ $merged_ty _varint >](buf) {
+          match $decode_fn(buf) {
             Ok((bytes_read, merged)) => {
               let (numer, denom) = $crate::utils::[< unpack_ $sign $bits >](merged);
               if denom == 0 {
@@ -78,9 +83,21 @@ macro_rules! impl_varint_for_ratio {
     }
   };
   (@const $($bits:literal($merged_ty:ident)),+$(,)?) => {
-    impl_varint_for_ratio!(@const_inner $(u::$bits($merged_ty)),+);
-    impl_varint_for_ratio!(@const_inner $(i::$bits($merged_ty)),+);
-  }
+    paste::paste! {
+      impl_varint_for_ratio!(@const_inner $(u::$bits($merged_ty) {
+        encoded_len_fn: $crate::[< encoded_ $merged_ty:snake _varint_len >],
+        encode_fn: $crate::[< encode_ $merged_ty:snake _varint >],
+        encode_to_fn: $crate::[< encode_ $merged_ty:snake _varint_to>],
+        decode_fn: $crate::[< decode_ $merged_ty:snake _varint >],
+      }),+);
+      impl_varint_for_ratio!(@const_inner $(i::$bits($merged_ty) {
+        encoded_len_fn: $crate::[< encoded_ $merged_ty:snake _varint_len >],
+        encode_fn: $crate::[< encode_ $merged_ty:snake _varint >],
+        encode_to_fn: $crate::[< encode_ $merged_ty:snake _varint_to>],
+        decode_fn: $crate::[< decode_ $merged_ty:snake _varint >],
+      }),+);
+    }
+  };
 }
 
 impl_varint_for_ratio!(8(u16), 16(u32), 32(u64), 64(u128));
@@ -91,6 +108,10 @@ mod ruint_1;
 
 #[cfg(feature = "bnum_0_13")]
 mod bnum_0_13;
+
+#[cfg(feature = "bnum_0_13")]
+#[cfg_attr(docsrs, doc(cfg(feature = "bnum_0_13")))]
+pub use bnum_0_13::*;
 
 #[cfg(test)]
 mod tests;